UPSTREAM: dma-fence: add comment for WARN_ON in dma_fence_release()
In dma_fence_release() there is a WARN_ON which could be triggered by
several cases of wrong dma-fence usage. This patch adds a comment to
explain two use-cases to help driver developers that use dma-fence
and trigger that WARN_ON to better understand the reasons for it.

v2: change to a more generic, one-liner comment

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Change-Id: I1a3447db798161f65f33fd019acfc0cec8c2dd60
Signed-off-by: Oded Gabbay <oded.gabbay@gmail.com>
Signed-off-by: nostalgiceagle <183935403+nostalgiceagle@users.noreply.github.com>